requirements.txt
Django==2.2.1
django-crispy-forms==1.7.2
django-filter==2.1.0
django-search-views==0.3.7
django-widget-tweaks==1.4.3
Pillow==6.0.0
pkg-resources==0.0.0
pytz==2019.1
sqlparse==0.3.0
